function_app/requirements.txt
azure-functions>=1.17.0
pydantic>=2.0,<3.0
